Support alternative environment implementations
This allows for adding an OpenStack environment type
