+++ vars/coreKubicProjectCi.groovy
@@ -28,6 +28,7 @@ def call(Map parameters = [:]) {
 
     withKubicEnvironment(
             nodeLabel: 'leap42.2&&m1.xlarge',
+            environmentType: 'devenv',
             gitBase: 'https://github.com/kubic-project',
             gitBranch: env.getEnvironment().get('CHANGE_TARGET', env.BRANCH_NAME),
             gitCredentialsId: 'github-token',
